message

+++ paraphrasing/my_data_handling.py
@@ -82,7 +82,7 @@ class DataHandler:
         batch['trg_ids'] = np.array([row + ([self.stoi['<PAD>']] * (max_len_t - len(row))) for row in batch['trg_ids']])
         return batch
     def iter_train_batches(self, batch_size):
-        self.train_total_batches = len(self.train_instances) / batch_size
+        self.train_total_batches = int(len(self.train_instances) / batch_size)
         pbar         = tqdm(total=self.train_total_batches+1)
         batch        = {'src_ids': [], 'trg_ids': []}
         for text_s, text_t in self.train_instances:
@@ -96,7 +96,7 @@ class DataHandler:
             pbar.update(1)
             yield self.fix_one_batch(batch)
     def iter_dev_batches(self, batch_size):
-        self.dev_total_batches   = len(self.dev_instances) / batch_size
+        self.dev_total_batches   = int(len(self.dev_instances) / batch_size)
         pbar                     = tqdm(total=self.dev_total_batches+1)
         batch                    = {'src_ids': [], 'trg_ids': []}
         for text_s, text_t in self.dev_instances:
